Refactor: 날자 표기 조건식을 Adapter에서 Fragment로 이동 #360
리스트 어댑터에서 계산하지 않도록 수정
