+++ agijagi/app/src/main/java/likelion/project/agijagi/notification/NotificationListFragment.kt
@@ -19,6 +19,8 @@ import likelion.project.agijagi.MainActivity
 import likelion.project.agijagi.R
 import likelion.project.agijagi.databinding.FragmentNotificationListBinding
 import likelion.project.agijagi.model.UserModel
+import java.text.SimpleDateFormat
+import java.util.Locale
 
 class NotificationListFragment : Fragment() {
 
@@ -193,6 +195,36 @@ class NotificationListFragment : Fragment() {
                     for (data in datas) {
                         val content = data.getString("content")!!
                         val date = data.getString("date")!!
+
+                        // 날자 계산 필요. pattern: yyMMddHHmmssSSS
+                        val tempDate = SimpleDateFormat("yyMMddHHmmssSSS").parse(date)
+                        val now = MainActivity.getMilliSec()
+                        val year = now.substring(0, 2).toInt() - date.substring(0, 2).toInt()
+                        val month = now.substring(2, 4).toInt() - date.substring(2, 4).toInt()
+                        val day = now.substring(4, 6).toInt() - date.substring(4, 6).toInt()
+
+                        var dateStr = ""
+                        if (1 < year) {
+                            // 1년이 지났다면 연도로 표기
+                            if (month < 1) {
+                                dateStr =
+                                    SimpleDateFormat("M월 d일", Locale.getDefault()).format(tempDate)
+                            } else {
+                                dateStr =
+                                    SimpleDateFormat("YYYY년", Locale.getDefault()).format(tempDate)
+                            }
+                        } else if (1 < day) {
+                            // 1년이 지나지 않았다면 00월 00일로 표기
+                            dateStr =
+                                SimpleDateFormat("M월 d일", Locale.getDefault()).format(tempDate)
+                        } else if (1 == day) {
+                            dateStr = "어제"
+                        } else {
+                            // 하루 이내의 시간은 오전/후 hh:mm 로 표기
+                            dateStr =
+                                SimpleDateFormat("aa h:m", Locale.getDefault()).format(tempDate)
+                        }
+
                         val is_chat = data.getBoolean("is_chat")!!
                         val is_read = data.getBoolean("is_read")!!
 
@@ -235,6 +267,7 @@ class NotificationListFragment : Fragment() {
                             senderName,
                             content,
                             date,
+                            dateStr,
                             type.str,
                             is_read
                         )
